+++ src/lib/utils/pinnedDocumentations/togglePinned.ts
@@ -36,7 +36,9 @@ const togglePinned = async (
       ...provider._documentations.filter(
         (documentation) => !documentation.isPinned
       ),
-    ];
+    ].filter(
+      (doc, index, self) => index === self.findIndex((d) => d.id === doc.id)
+    );
 
     provider._view.webview.postMessage({
       type: "setDocumentations",